retain Agents.jl-like stepping functions
