+++ MainDummy.jl
@@ -16,7 +16,7 @@ julia> include("this_script.jl")
 using  SocialABMs: SocialABM, population_step!
 using  SocialSimulations: DummyExample
 using  SocialSimulations: ABMSocialSimulation
-using  SocialSimulations: dummystep, run!, attach_model_step! 
+using  SocialSimulations: run!, attach_model_step! 
 import SocialSimulations: setup!
 using Dummy: createPopulation # , loadData!
